# swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Adventure APIs",
    "description": "API for admin/managing and client/viewing",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/user/sign_up": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "user signup",
        "operationId": "user_signup",
        "parameters": [
          {
            "name": "phone_number_email",
            "in": "formData",
            "description": "phone number or email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "first_name",
            "in": "formData",
            "description": "first_name",
            "required": false,
            "type": "string"
          },
          {
            "name": "last_name",
            "in": "formData",
            "description": "last_name",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Email or phone_number is existed; -2: DB error; -3: Email or phone_number is required; -4 : Email is existed; -5 : Phone number is existed  ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "email": {
                      "type": "string",
                      "example": "Brucelee Thanh@gmail.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": ""
                    },
                    "gender": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "date-time",
                      "example": "1993-07-05T04:27:58.144Z"
                    },
                    "address": {
                      "type": "string",
                      "example": ""
                    },
                    "religion": {
                      "type": "string",
                      "example": ""
                    },
                    "intro": {
                      "type": "string",
                      "example": ""
                    },
                    "fb_id": {
                      "type": "string",
                      "example": ""
                    },
                    "avatar": {
                      "type": "string",
                      "example": ""
                    },
                    "cover": {
                      "type": "string",
                      "example": ""
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "last_visited_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "email is existed"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "user login",
        "operationId": "user_login",
        "parameters": [
          {
            "name": "phone_number_email",
            "in": "formData",
            "description": "phone number or email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "fcm_token",
            "in": "formData",
            "description": "fcm_token",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Email or phone_number is not existed; -2: DB error; -3: Email or phone_number is required;  ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "email": {
                      "type": "string",
                      "example": "Brucelee Thanh@gmail.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": ""
                    },
                    "gender": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "date-time",
                      "example": "1993-07-05T04:27:58.144Z"
                    },
                    "address": {
                      "type": "string",
                      "example": ""
                    },
                    "religion": {
                      "type": "string",
                      "example": ""
                    },
                    "intro": {
                      "type": "string",
                      "example": ""
                    },
                    "fb_id": {
                      "type": "string",
                      "example": ""
                    },
                    "avatar": {
                      "type": "string",
                      "example": ""
                    },
                    "cover": {
                      "type": "string",
                      "example": ""
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "last_visited_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "token": {
                      "type": "string",
                      "example": "923749hwhff2rhfwhf-fhfhdkwehf-fhdfjdfhd"
                    },
                    "fcm_token": {
                      "type": "string",
                      "example": "923749hwhff2rhfwhf-fhfhdkwehf-fhdfjdfhd"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "email is existed"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/fb_login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "user fb login",
        "operationId": "fb_login",
        "parameters": [
          {
            "name": "fb_token",
            "in": "formData",
            "description": "fb_token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -2: DB error; -3: Account is existed  ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Brucelee Thanh"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "email": {
                      "type": "string",
                      "example": "Brucelee Thanh@gmail.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": ""
                    },
                    "gender": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "date-time",
                      "example": "1993-07-05T04:27:58.144Z"
                    },
                    "address": {
                      "type": "string",
                      "example": ""
                    },
                    "religion": {
                      "type": "string",
                      "example": ""
                    },
                    "intro": {
                      "type": "string",
                      "example": ""
                    },
                    "fb_id": {
                      "type": "string",
                      "example": "fw23fwefw23r2rwfwwf"
                    },
                    "avatar": {
                      "type": "string",
                      "example": ""
                    },
                    "cover": {
                      "type": "string",
                      "example": ""
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "last_visited_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "token": {
                      "type": "string",
                      "example": "923749hwhff2rhfwhf-fhfhdkwehf-fhdfjdfhd"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "email is existed"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}